requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aisuite==0.1.10
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
courlan==1.3.2
dateparser==1.2.1
debugpy==1.8.12
decorator==5.2.0
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.7.4
fastapi==0.115.8
filelock==3.17.0
fissa==1.0.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
h11==0.14.0
htmldate==1.9.3
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jusText==3.0.1
lazy_loader==0.4
lxml==5.3.1
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.3
openai==1.63.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
read-roi==1.6.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==3.4.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.2.18
tld==0.13
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
yarl==1.18.3
